move CodableFeedStore to its own file
